requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
PyMySQL==1.1.0
python-dotenv==1.0.0
pdfkit==1.0.0
gunicorn==21.2.0
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
Bootstrap-Flask==2.2.0
Werkzeug==3.1.4
Jinja2==3.1.6


# Testing Dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.2.0
factory-boy==3.3.0